P1/requirements.txt
ecdsa
pycryptodomex
scipy
